Fix cast in 0-binary_to_uint

+++ 0x14-bit_manipulation/0-binary_to_uint.c
@@ -17,7 +17,7 @@ unsigned int binary_to_uint(const char *b)
 		return (0);
 	while (b[size] != '\0')
 		size++;
-	if (size >= (sizeof(unsigned int) * 8))
+	if (size >= ((int)(sizeof(unsigned int) * 8)))
 		return (0);
 	for (i = 0; i < size; i++)
 	{
